+++ CHANGELOG.md
@@ -6,3 +6,5 @@ the recommendations of [keepachangelog.com](http://keepachangelog.com/).
 ## 1.0.0
 
 - [PT #1070](https://github.com/paper-trail-gem/paper_trail/issues/1070), [#2](https://github.com/westonganger/paper_trail-association_tracking/issues/2) - Extracted from paper_trail gem in v9.2
+
+- [PT #594](https://github.com/paper-trail-gem/paper_trail/issues/594), Fix for issues around reify with models using STI.